% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-bangalore.R
\docType{data}
\name{bangalore_boundary}
\alias{bangalore_boundary}
\title{Bangalore Morphology Data}
\format{
\verb{sf POLYGON} geometries that define the extents of
the study area.
}
\usage{
bangalore_boundary
}
\description{
These data sets contain example morphometric data for Bangalore and were
obtained from OpenStreetMap using the \code{osmdata} package.
Bangalore Boundary
}
\keyword{datasets}
