% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-bangalore.R
\docType{data}
\name{bangalore_buildings}
\alias{bangalore_buildings}
\title{Bangalore Buildings}
\format{
\verb{sf POLYGON} geometries of buildings in the study
area.
}
\usage{
bangalore_buildings
}
\description{
Bangalore Buildings
}
\keyword{datasets}
