% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi.R
\name{cc}
\alias{cc}
\title{Canonical Correlation Analysis}
\usage{
cc(X, Y)
}
\arguments{
\item{X}{is a 1st Group}

\item{Y}{is a 2nd Group}
}
\description{
The Discrimination & Classification are multivariate techniques concered with separating distint sets of objects (or observations) and with allocating new objects (observations) to perviously define groups.
}
\references{
Inspired by Gold Medalist Dr Shahla Faisal who study me in GCUF
}
\author{
Syed Hammad Raza (Studies BS Statistics in GCUF)
}
