% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi.R
\name{ccs}
\alias{ccs}
\title{Canonical Correlation of Covariance or Correlation Matrix}
\usage{
ccs(sigma)
}
\arguments{
\item{sigma}{must ba a covariance or correlation matrix}
}
\description{
The Discrimination & Classification are multivariate techniques concered with separating distint sets of objects (or observations) and with allocating new objects (observations) to perviously define groups.
}
\references{
Inspired by Gold Medalist Dr Shahla Faisal who study me in GCUF
}
\author{
Syed Hammad Raza (Studies BS Statistics in GCUF)
}
