% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_create_proximitybands.R
\name{st_create_proximitybands}
\alias{st_create_proximitybands}
\title{Create proximity band spatial units.}
\usage{
st_create_proximitybands(
  x,
  band_width = 10,
  segment_length = 2,
  merge_threshold = 1015,
  merge_type = "min_centroid_distance",
  contiguity = "rook",
  verbose = T
)
}
\arguments{
\item{x}{an \code{sf} object with \code{LINESTRING} geometries representing a street
network.}

\item{band_width}{numeric; width of the band as measured from the centerline
of the \code{LINESTRING} representing the street.}

\item{segment_length}{numeric; Length of segments street linestrings are
subdivided into. Default = 2.}

\item{merge_threshold}{numeric; value represents the smallest acceptable
size for a spatial unit. Contiguous units will be iteratively merged until
this value is reached. To skip this process, set \code{merge_threshold = NULL}.
See st_merge_spatialunits() for more details.}

\item{merge_type}{string; Passed on to st_merge_spatialunits().
Criteria with which polygons are merged. Must be one of
"min_centroid_distance", "min_shared_boundary", or "max_shared_boundary".
Default = "min_centroid_distance".
See st_merge_spatialunits() for more details.}

\item{contiguity}{string; one of "queen" or "rook". Default = "rook".}

\item{verbose}{logical; if \code{FALSE} no status messages will be output.}
}
\value{
An \code{sf} object with \code{POLYGON} geometries representing tessellated
spatial units.
}
\description{
Create proximity band spatial units.
}
\examples{
proximity_bands <- st_create_proximitybands(x = bangalore_highways,
verbose = FALSE, merge_threshold = NULL)
plot(proximity_bands)
}
\references{
Araldi, A., & Fusco, G. (2019). From the street to the
metropolitan region: Pedestrian perspective in urban fabric analysis.
Environment and Planning B: Urban Analytics and City Science, 46(7),
1243–1263. https://doi.org/10.1177/2399808319832612
}
\seealso{
st_merge_spatialunits, \code{\link[moter:motess]{moter::motess()}}
}
