% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_explode.R
\name{st_explode}
\alias{st_explode}
\title{Explode geometry to manipulate constituent elements.}
\usage{
st_explode(x)
}
\arguments{
\item{x}{an \code{sf} object with \code{LINESTRING} geometries.}
}
\value{
An \code{sf} object with \code{LINESTRING} geometries with each line segment in
a separate row.
}
\description{
Explode geometry to manipulate constituent elements.
}
\examples{
exploded_geom <- st_explode(bangalore_highways)
exploded_geom
}
