% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_extract_connected.R
\name{st_extract_connected}
\alias{st_extract_connected}
\title{Extract connected streets from a data frame containing linestrings representing street networks.}
\usage{
st_extract_connected(x)
}
\arguments{
\item{x}{a \code{sf} object with \code{LINESTRING} geometries.}
}
\value{
A \code{sf} object containing a subset of \code{x}.
}
\description{
Extract connected streets from a data frame containing linestrings representing street networks.
}
\examples{
connected_streets <- st_extract_connected(bangalore_highways)
connected_streets
}
