% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_extract_connected.R
\name{st_extract_disconnected}
\alias{st_extract_disconnected}
\title{Extract connected streets from a data frame containing linestrings representing street networks.}
\usage{
st_extract_disconnected(x)
}
\arguments{
\item{x}{an \code{sf} object with \code{LINESTRING} geometries.}
}
\value{
An \code{sf} object containing a subset of \code{x}.
}
\description{
Extract connected streets from a data frame containing linestrings representing street networks.
}
\examples{
disconnected_streets <- st_extract_disconnected(bangalore_highways)
disconnected_streets
}
