% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_extract_intersections.R
\name{st_extract_intersections}
\alias{st_extract_intersections}
\title{Extract street intersections from a data frame containing linestrings representing street networks.}
\usage{
st_extract_intersections(x)
}
\arguments{
\item{x}{an \code{sf} object with \code{LINESTRING} geometries.}
}
\value{
An \code{sf} object with \code{POINT} geometries representing street intersections.
}
\description{
Extract street intersections from a data frame containing linestrings representing street networks.
}
