% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_merge_spatial_units.R
\name{st_merge_spatial_units}
\alias{st_merge_spatial_units}
\title{Title}
\usage{
st_merge_spatial_units(x, merge_threshold, verbose = T)
}
\arguments{
\item{x}{an \code{sf} object with \code{POLYGON} geometries representing the
spatial units to be merged.}

\item{merge_threshold}{numeric; value represents the smallest acceptable
size for a spatial unit. Contiguous units will be iteratively merged until
this value is reached. To skip this process, set \code{merge_threshold = NULL}.}

\item{verbose}{logical; if \code{FALSE} no status messages will be output.}
}
\value{
An \code{sf} object with \code{POLYGON} geometries representing street block
spatial units.
}
\description{
Title
}
