% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_merge_spatialunits.R
\name{st_merge_spatialunits}
\alias{st_merge_spatialunits}
\title{Iteratively merge spatial units to ensure all of them meet a minimum area threshold.}
\usage{
st_merge_spatialunits(x, merge_threshold, verbose = T)
}
\arguments{
\item{x}{an \code{sf} object with \code{POLYGON} geometries representing the
spatial units to be merged.}

\item{merge_threshold}{numeric; value in square meters represents the
smallest acceptable area for a spatial unit. Contiguous units will be
iteratively merged until this value is reached. To skip this process,
set \code{merge_threshold = NULL}.}

\item{verbose}{logical; if \code{FALSE} no status messages will be output.}
}
\value{
An \code{sf} object with \code{POLYGON} geometries representing street block
spatial units.
}
\description{
Iteratively merge spatial units to ensure all of them meet a minimum area threshold.
}
