% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_orientation.R
\name{st_orientation}
\alias{st_orientation}
\title{Determine the interior angle between a \code{LINESTRING} and the horizontal.}
\usage{
st_orientation(x)
}
\arguments{
\item{x}{a \code{sf} object containing \code{LINESTRING} geometries.}
}
\value{
An unnested \code{sf} object containing the orientation of each line
segment making up the original \code{LINESTRING} geometries.
}
\description{
Determine the interior angle between a \code{LINESTRING} and the horizontal.
}
