% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_survey_spatialunits.R
\name{st_survey_spatialunits}
\alias{st_survey_spatialunits}
\title{Survey a set of spatial units to get summary statistics.}
\usage{
st_survey_spatialunits(x, trim = NULL)
}
\arguments{
\item{x}{an \code{sf} object with \code{POLYGON} geometries representing the
spatial units to be surveyed.}

\item{trim}{numeric; fraction (0 to 0.5) of values to be trimmed from each
end to compute a trimmed mean}
}
\value{
A \code{data.frame} object with numeric values of min, max, mean, and
median.
}
\description{
Survey a set of spatial units to get summary statistics.
}
\examples{
street_blocks <- st_create_streetblocks(x = bangalore_highways,
boundary = bangalore_boundary, merge_threshold = NULL, verbose = FALSE)
survey_report <- st_survey_spatialunits(x = street_blocks)
survey_report
}
