% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_nest_coordinates.R
\name{st_unnest_coordinates}
\alias{st_unnest_coordinates}
\title{Unnest \code{POINT} geometries from \code{LINESTRINGS}}
\usage{
st_unnest_coordinates(x)
}
\arguments{
\item{x}{a \code{sf} object containing \code{LINESTRING} geometries.}
}
\value{
A \code{df} where each row contains the \code{x1}, \code{y1}, \code{x2}, and \code{y2}
coordinates for a line.
}
\description{
Unnest \code{POINT} geometries from \code{LINESTRINGS}
}
