% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-bangalore.R
\docType{data}
\name{bangalore_boundary}
\alias{bangalore_boundary}
\title{Bangalore Boundary}
\format{
\verb{sf POLYGON} geometries that define the extents of
the study area.
}
\usage{
bangalore_boundary
}
\description{
This is one of three datasets with example morpometric data obtained from
OpenStreetMap using the \code{osmdata} package.
Data (c) OpenStreetMap contributors, ODbL 1.0.
https://www.openstreetmap.org/copyright
This dataset contains geometry that defines the extents of the study area
in Bangalore.
}
\keyword{datasets}
