% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-bangalore.R
\docType{data}
\name{bangalore_buildings}
\alias{bangalore_buildings}
\title{Bangalore Buildings}
\format{
\verb{sf POLYGON} geometries of buildings in the study
area.
}
\usage{
bangalore_buildings
}
\description{
This is one of three datasets with example morpometric data obtained from
OpenStreetMap using the \code{osmdata} package.
Data (c) OpenStreetMap contributors, ODbL 1.0.
https://www.openstreetmap.org/copyright
This dataset contains geometry that defines the buildings in the study area
in Bangalore.
}
\keyword{datasets}
