% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-bangalore.R
\docType{data}
\name{bangalore_highways}
\alias{bangalore_highways}
\title{Bangalore Highways}
\format{
\verb{sf LINESTRING} geometries of streets in the study
area.
}
\usage{
bangalore_highways
}
\description{
Bangalore Highways
}
\keyword{datasets}
