% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_create_streetblocks.R
\name{st_create_streetblocks}
\alias{st_create_streetblocks}
\title{Create street block spatial units.}
\usage{
st_create_streetblocks(
  x,
  boundary,
  merge_threshold = 4050,
  merge_type = "min_centroid_distance",
  contiguity = "rook",
  verbose = T
)
}
\arguments{
\item{x}{a \code{sf} object with \code{LINESTRING} geometries representing a street
network.}

\item{boundary}{an \code{sf} object with \code{POLYGON} geometries representing the
extents of the study area.}

\item{merge_threshold}{numeric; value represents the smallest acceptable
size for a spatial unit. Contiguous units will be iteratively merged until
this value is reached. To skip this process, set \code{merge_threshold = NULL}.
See st_merge_spatialunits() for more details.}

\item{merge_type}{string; Passed on to st_merge_spatialunits.
Criteria with which polygons are merged. Must be one of
"min_centroid_distance", "min_shared_boundary", or "max_shared_boundary".
Default = "min_centroid_distance".
See st_merge_spatialunits() for more details.}

\item{contiguity}{string; one of "queen" or "rook". Default = "rook".}

\item{verbose}{logical; if \code{FALSE} no status messages will be output.}
}
\value{
A \code{sf} object with \code{POLYGON} geometries representing street block
spatial units.
}
\description{
Create street block spatial units.
}
\examples{
street_blocks <- st_create_streetblocks(x = bangalore_highways,
boundary = bangalore_boundary, merge_threshold = 1, verbose = FALSE)
plot(street_blocks)
}
\seealso{
st_merge_spatialunits()
}
