% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersections.R
\name{st_extract_intersections}
\alias{st_extract_intersections}
\title{Extract street intersections.}
\usage{
st_extract_intersections(x)
}
\arguments{
\item{x}{an \code{sf} object with \code{LINESTRING} geometries representing street networks.}
}
\value{
An \code{sf} object with \code{POINT} geometries representing street intersections.
}
\description{
Extract street intersections.
}
\examples{
intersections <- st_extract_intersections(bangalore_highways)
plot(bangalore_highways, col = "black")
plot(intersections, col = "red", add = TRUE)
}
